Update patch set 6
Patch Set 6:

As I dig more into converting more objects to use pools, I'm having conversations with others that are not so much fans of the counters.

I'm wondering if there are real situations you have needed to detect memory leaks where valgrind massif wouldn't be just as effective. I get that a customer environment might not be a place we can run valgrind massif, and there is something to be said for the counters that are just always there to be queried, but how much is this need?

Can we limit the pool counting to just certain objects? I was starting to go hog wild converting everything, but maybe that's overkill.

Patch-set: 6
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>