NFS: properly handle NULL fs_locations pointer
RFC5661, section 11.9:

   When the fs_locations attribute is interrogated and there are no
   alternate file system locations, the server SHOULD return a zero-
   length array of fs_location4 structures, together with a valid
   fs_root.

If we have a NULL fs_locations pointer, we currently fill in a
structure with a bunch of "not supported" strings. Change the code
to handle it more in line with what the spec suggests.

There is one wrinkle here, which is that if we don't have an
fs_locations structure, then we also don't have an fs_root. In that
case, just assume that the entire export has gone away and set the
fs_root to the pseudopath of the export.

Change-Id: I11e4db806dcbf3c853b066b5a97f950c57ec66c1
Signed-off-by: Jeff Layton <jlayton@redhat.com>