FSAL: make fsal_fs_locations_t.server an array of hosts
Add a flexarray at the end of fsal_fs_locations_t to hold an array of
utf8string structs. For now, we assume that the rootpath on all of them
is the same.

Change nfs4_fs_locations_new to take a server count and allocate a
structure that's the right length for the array. The callers then
populate the array with the appropriate info, which is then passed
directly to the encoder.

Change-Id: Iffdc6e90aacbce58eefe066d06641af07f370553
Signed-off-by: Jeff Layton <jlayton@redhat.com>