Update patch set 2
Patch Set 2:

> > Patch Set 2:
 > >
 > > > When testing the latest code, I get many complaints at nfs
 > client
 > >  > as,
 > >  >
 > >  > kernel: NFS: directory fsstress.48865.2/p0 contains a readdir
 > >  > loop.Please contact your server vendor.  The file: d3XXXXXXXX
 > has
 > >  > duplicate cookie 2
 > >  > kernel: NFS: directory p0/d3XXXXXXXX contains a readdir
 > loop.Please
 > >  > contact your server vendor.  The file: dbX has duplicate
 > cookie 2
 > >
 > > What file system are you using under FSAL_VFS? How does your file
 > system produce the cookies (struct dirent d_off field)? If it can
 > generate 0, 1, or 2 that may be an issue.
 > 
 > Ganesha currently assumes it can use 0, 1, and 2 as reserved
 > values.  This was always fine before, but I wonder if this was
 > because the first dirent had an offset of 3, and ".." had the
 > offset of 2, but now we're using the previous d_off, which is now
 > 2?

Hmm, FSAL_VFS should start with cookie 3 if the readdir starts from the beginning, see how it sets cookie (from whence if whence != 0 or FIRST_COOKIE if whence == 0). nfs3_readdirplus and nfs3_readdir set fsal_cookie to begin_cookie if begin_cookie > 2 otherwise to 0 (because fsal_readdir doesn't enumerate "." and "..") but NFSv3 READDIR DOES enumerate "." and ".." (while NFSv4 READDIR DOES NOT enumerate "." and "..").

So as far as I can tell, FSAL_VFS read_dirents should never make a callback with cookie 2 unless the underlying file system actually enumerates an entry where d_off is 2. I guess if it does enumerate a ".." entry with a d_off of 2 that could happen, but that implies the next dirent after ".." would be found by doing a seekdir to 2.

I guess that could happen if FSAL_VFS for example was re-exporting an NFS share.

Maybe there are other file systems that do enumerate "." with cookie 1 and ".." with cookie 2.

It may be that on skipping "." and ".." we should check what d_off is, and if it's 2, set cookie to 3 and continue instead of goto skip. I'm not sure what to do if a file system enumerates in a way that "." and ".." are not enumerated immediately in sequence but are enumerated with cookie 1 and cookie 2. EXT4 enumerates "." and ".." somewhere in the middle of the directory, but their d_off is the hash of the "." or ".." string and EXT4 enumerates in hash order.

Patch-set: 2
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>