Update patch set 2
Patch Set 2:

Sorry for the false start, the basic idea was sound, but it had some race conditions and was the wrong place to fix this. This is what we really need to do in this situation. Ensure that we don't fully lift the grace period until all the cluster nodes are enforcing.

Patch-set: 2
Reviewer: Jeff Layton <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>