Update patch set 2
Patch Set 2:

I tried creating an f2fs filesystem with a bunch of files. I see d_off of 1 for "." and 2 for "..", but then subsequent d_off are 4, 6, 8, ...
So using 3 for the first entry would not cause collision.

I'm not sure if there are different ways to create the filesystem or if a larger directory would result in a d_off of 3.

My filesystem is on a 120 MB file created with default options.

Patch-set: 2
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>