Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Something is quite wrong here.  op_ctx is set in mdc_up_lock_grant() and mdc_up_lock_avail(), so no async lock grants should have op_ctx unset.  Can you be more specific about how this happens, with a backtrace?
> 
> -1 until we work this out.

Okay, so I saw the backtrace in the email.  The problem appears to be that SAL has it's own async thread that runs on top of UP's async thread, losing the work that MDCACHE did to set up op_ctx.   So something like this is necessary, but it should probably go in either state_blocked_lock_caller() or process_blocked_lock_upcall(), since they're the top of the stack for the new thread, and in general we assume (especially in MDCACHE) that op_ctx is valid.

Patch-set: 1
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>