FSAL: eliminate fsal_grace boolean
The fsal_staticfsinfo_t struct has a boolean called fsal_grace that will
offload grace period handling to the FSAL. Nothing defaults to this
being on, but GPFS has an option that allows it to be enabled.

Removing this boolean simplifies the code significantly, and it's not
clear that it's actually being used. This patch removes it, under the
assumption that it's no longer useful.

Cc: Malahal Naineni <malahal@gmail.com>
Change-Id: Ic4af1f814e86175553a3a57eedc771b17947dee2
Signed-off-by: Jeff Layton <jlayton@redhat.com>