Update patch set 1
Patch Set 1:

Regardless, it would be good to have a clear description of what the semantics are when a FSAL sets this to true.

For instance, in the NLM code we only check for fsal_grace if nfs_in_grace returns true. Is it possible that nfs_in_grace could return false when the grace period is actually active? Is that a problem here? Why check anything at all in ganesha if you're setting this field to true?

A lot of the logic around this setting is unclear, and I think there are some potential races here that could lead to subtle recovery bugs.

Patch-set: 1
Reviewer: Jeff Layton <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>