Update patch set 2
Patch Set 2:

I have pointed people interested in a bugzilla with respect to using BTRFS under Ganesha here...

As I try to understand all the implications going on here, I'm starting to think this patch is a bad idea.

I had this patch idea hanging around as an FYI patch to help with an FYI demonstration of how you could implement compute_readdir_cookie in FSAL_VFS. When some directory testing on FSAL_GLUSTER seemed to show different directory entries having the same telldir cookie, I thought this idea would solve the issue. Unfortunately, the issue is caused by other implication of how readdir works on Gluster and I no longer think this is a useful solution.

In theory, this solution had the potential of stabilizing cookies over directory mutation, but there are still things that can happen that destabilize the dirent cache, and in reality, directory mutation should result in the directory being invalidated so the benefits of this patch for stabilizing cookies across directory mutation are less important.

This patch introduced a problem of what cookie to use for the first directory entry.

BTRFS emits the following directory entries upon starting to enumerate the directory:

name = ".", d_off = 1
name = "..", d_off = 2
name = "foo", d_off = 3
name = "bar", d_off = 4

That is assuming foo was the first file created in the directory (BTRFS actually has stable cookies under directory mutation because a dirent is assigned a cookie from a monotonically increasing counter as files are created). The problem is that we can't use 1 or 2 as a cookie, so if the cookie from the previous entry was 1 or 2, we were assigning 3 to "foo", but that was going to be the cookie for "bar", so we have a duplicate cookie problem.

In contrast I have observed small directories in EXT4 having a pattern like

name = "foo", d_off = 0x06ec1f7d0a07b020
name = "bar", d_off = 0x094bb60c7458c6b0

and then sometime later we see:

name = "baz", d_off = 0x4c470ee8300a65ab
name = ".", d_off = 0x5660f9c91e0d9a72

and then sometime later:

name = "bozo", d_off = 0x68ec4bc2e1982399
name = "..", d_off = 0x6ff6e2f3ce892cbb

Note that the entries for the dirent previous to "." or ".." seem to have a stable cookie value suggesting that 0x06ec1f7d0a07b020 is the hash of "." and 0x68ec4bc2e1982399 is the hash of "..".

The problem is what previous d_off should we use as the cookie for the first entry "foo"? We can't use 0 (since NFS reserves that value). So again, we used 3 with perhaps less risk of a collision (something would have to hash to 3).

I am now thinking we should just return an entries d_off as it's cookie and rely on invalidation to occur if the directory is mutated. And even if the directory mutates without us detecting a change and invalidating, we can never have a perfect solution (except maybe on BTRFS...) so let's not try.

The problem Gluster was having that seemed to be solved (or at least improved) with this similar change in Gluster needs to be solved a different way.

Patch-set: 2
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>