Revert "VFS: Use d_off from previous dirent as cookie"
This has been demonstrate to no longer be a good idea.

This is not a strict revert as we needed to account for FreeBSD
changes made after the reverted patch. To maintain bisectability
those adjustments are included in this revert.

Basically on FreeBSD, we need to do extra work to get the cookie
for each dirent.

This reverts commit 78cc090028f0fe10a7a73de640e52f5d6456a705.

Change-Id: I6a7e22726ff1e95b2a945588d8077189e8f551ba