Update patch set 1
Patch Set 1:

> Right, kernel won't hardcode . and .. to 1/2, but I believe it
 > should not give 1/2 to any entry either... Hard to tell but I don't
 > see anything in the nfsd code to protect the values, and they have
 > a similar check than us in the readdir nfs4 op (return bad cookie
 > if 1 or 2 comes back)
 > 
 > The thing is, we weren't returning the offset of the last entry but
 > one-to-last, the example program in getdents man page was showing
 > d_off == 2 for the file through 9p mount but it's 3 on the server;
 > this appears fixed now so I'll just drop this patch.

Yea, I was attempting to solve a problem with dirent cookies and volatile directories by using the cookie from the previous dirent. This worked fine for ext4 (which, for smaller directories at least, uses the hash of the name "." or ".." as the seekdir offset that would find "." or ".." - ext4 also does not return "." and ".." as the 1st and 2nd directory entries, do an ls -aU to observe this, you will see different cookies for "." and ".." if you write a program to call getdents/readdir and print name and d_off, but you will notice the entries before always have consistent cookies).

But this technique caused problems with filesystems that used 1 and 2 for "." and ".." and a simple fix to not return a cookie smaller than 3 caused problems with btrfs for which the first file created in a directory will have cookie 3 (i.e. a seekdir(2) on btrfs will find that first file and it's d_off will be 3).

In the end, there are still directory mutation situations that are not properly solved by using the previous cookie, so I've reverted the patch, and thus cookies 1 and 2 will never be used (and a first btrfs file will indeed have cookie 3).

I still have no idea if ext4 assures that it's hashing doesn't produce 0, 1, or 0x7fffffffffffffff (the cookie it uses for the last directory entry - if a file hashed to that, it WOULD be the last file, but the penultimate file would have cookie 0x7fffffffffffffff also...). But with 64 bit (well, actually 63 bit) hash, production of any specific hash result is extremely unlikely. Actually, in the case of a file name hashing to 0x7fffffffffffffff ext4 would probably just emit two entries with that cookie...

Patch-set: 1