MDCACHE - Keep dir handles open across readdir
Technically, there's no correlation between offsets returned by
telldir(3) during one opendir(3) and another opendir(3).  However, this
happens to work on most filesystems, but not all.  To fix this, open dir
handles when the first dirent is read from the FSAL, and leave them open
until either 1) the readdir completes on that director; or 2) the LRU
demotes the handle and closes it.

Change-Id: I5850dd81523ab741c0dc659d7511c8c358d34740
Signed-off-by: Daniel Gryniewicz <dang@redhat.com>