Update patch set 1
Patch Set 1:

> > Patch Set 1:
 > >
 > > What happens if the user does an open(READ), lock(READ), then
 > permissions are changed to make the file read-only, then the user
 > does open(WRITE), lock(WRITE)?  The open(WRITE) should fail,
 > because of new permissions, but it will succeed because we already
 > had the file open RDWR.
 > 
 > Wouldn't the open(WRITE) fail at that point? open_for_locks would
 > be false when you called into fsal_find_fd for the open call, no?

This is only NFSv4 (NLM can not have associated openstate). In that case, the OPEN should actually be preceded by an ACCESS, but even if not (and it really has to, be cause we allow OPEN of an execute only file, that checking MUST be done by ACCESS), permissions will be checked on that OPEN in the protocol layer before calling FSAL open2. And when the FSAL open2 is made, that is for the openstate and thus open_for_locks is false. Note that for NFS v3 we ALWAYS open RDWR because we don't know what the actual open system call was for (NFSv3 not having an OPEN procedure).

There really is no other way to handle this re-open situation that I can see.

Patch-set: 1
Reviewer: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>