FSAL_MDCACHE: prevent double free on new raced entries
This is a partial revert of c55046feb786d69de8ba046e7cbd242479621b66.
The newly acquired entry is actually in the LRU at this point,
it is inserted in mdcache_alloc_handle.
We should put the ref once and let the reaper do the rest,
otherwise it may race with another thread (LRU thread/reaper).

Change-Id: Ia548c842b79958deb6b895a0249ec0bb4cde65d2
Signed-off-by: Fatih Acar <fatih.acar@gandi.net>