FSAL_MDCACHE: prevent double free on new raced entries
The newly acquired entry is actually in the LRU when trying to double-put
references in order to free it. It is inserted in mdcache_alloc_handle.
We should do the insertion once we are sure the new entry is not raced
otherwise we may race with another thread (LRU thread/reaper).
Potential case is when a long delay happens between the double-put,
the entry may get reaped and cleaned before the last ref is put.

Change-Id: Ia548c842b79958deb6b895a0249ec0bb4cde65d2
Signed-off-by: Fatih Acar <fatih.acar@gandi.net>