namespace cleanup: some poorly named and/or non-static symbols
In Gluster are seeing a collision between the variable named 'options'
in the gluster xlators and the variable named 'options' in ganesha.nfsd.

ISTR a recent change made in Fedora to RT linker that affect this but
am unable to find a reference to the discussion at this time. It is
unclear to me whether this is a side effect of the introduction of the
-z defs linker flags in Fedora 28. This probably affects RHEL 8 too.

In a nutshell, references within a gluster xlator .so used to always be
resolved to the "closest" definition and/or resolved at static link time;
either way to the one in the xlator .so.

Now though we see them resolved at runtime to the first definition of
the symbol; in this case the one in ganesha.nfsd.

In addition to 'options' there are a number of other variables in
ganesha.nfsd that a) should be static, or b) should be renamed so as
not to pollute the namespace. I picked a few that stood out. There may
be others.

Note too that gluster is working to clean up its own scribbling in the
namespace. And, e.g., 'options' are exported from xlator .so files
so that each xlators' options can be enumerated for display.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: I0a5893ac4d54fce84e71f38997e1370f39d42ad3