namespace cleanup: some poorly named and/or non-static symbols
In Gluster are seeing a collision between the variable named 'options'
in the gluster xlators and the variable named 'options' in ganesha.nfsd.

(Fedora tried to introduce  -z defs to the linker flags, but has
ultimately decided not to make the change. It's not clear to mea that
this would change anything.)

In addition to 'options' there are a few other variables in ganesha.nfsd
that could be local in main() or should have a prefix added.

Note too that gluster is working to clean up its own scribbling in the
namespace. And, e.g., 'options' are exported from xlator .so files
so that each xlators' options can be enumerated for display.

Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>
Change-Id: I0a5893ac4d54fce84e71f38997e1370f39d42ad3