Update patch set 1
Patch Set 1:

So, I'm not sure about this, in principle.  By definition, Ganesha owns the namespace, since it's the application, and it's up to the libraries to not pollute it.

Are these symbols actually colliding with anything?  Or is this just a pre-emptive strike?

Patch-set: 1
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>