Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> So, I'm not sure about this, in principle.  By definition, Ganesha owns the namespace, since it's the application, and it's up to the libraries to not pollute it.
> 
> Are these symbols actually colliding with anything?  Or is this just a pre-emptive strike?

'options' is a collision now. Yes, ganesha owns the namespace, but it's 'friendlier' IMO if ganesha doesn't scribble indiscriminately in the namespace and (consistently) uses a prefix.

And yes, it's bad that gluster scribbles indiscriminately too, and we're working on fixing that, but it's not going to be fixed overnight, and probably not going to be backported to previous versions. :-/

Patch-set: 1
Reviewer: kaleb@redhat.com <1004389@d5d70762-12d0-45a1-890d-524b12d3f735>