Free group name string properly in proc_export()
If there is a client block with CIDR spec but the last client block is
not a CIDR spec, then the code will try to free unallocated memory
leading to a crash! Also, there is a memory leak if there are multiple
client blocks with CIDR spec.

Each client block with CIDR spec needs to free group name string.

Change-Id: I46d4a78d7a6a88f7374e547d59549f098cf1dfcd
Signed-off-by: Malahal Naineni <malahal@us.ibm.com>