SAL: convert current_grace to struct timespec
This allows us to also switch to using CLOCK_MONOTONIC, which guards
against clock jumps and skew. We really do want to base the grace
period on "absolute" time and not be affected by someone monkeying with
the host's clock.

This means that we need to take the grace_mutex a bit earlier when
attempting to lift the grace period, but that also allows us to
simplify nfs4_lift_grace_locked since we can assume that the mutex
was held over the timeout check.

Change-Id: I837bab37c447a9cb9e114b47cd25de8344c8609c
Signed-off-by: Jeff Layton <jlayton@redhat.com>