NFS: clean up and simplify RECLAIM_COMPLETE handling
We're currently tracking whether a client in the reclaim hash sent a
reclaim complete. We already track whether a client sent a
RECLAIM_COMPELTE though in the in-core clientid structure. That one is
more correct anyway, as the client could re-establish its clientid
during the grace period and attempt to re-perform reclaim.

Drop the flag from the reclaim clid table hash object and the function
that sets it. Move the increment of the reclaim_completes into
nfs4_op_reclaim_complete

Change-Id: I73cf293f9e26475b5a51bf41c9251ad44faf6813
Signed-off-by: Jeff Layton <jlayton@redhat.com>