RFC: [do not merge] (lib)ganesha_nfsd as a DSO
Build ganesha as a DSO, i.e. libganesha_nfsd.so.

The ganesha.nfsd daemon consists of just nfs_main::main() and is
linked with the DSO.

Then all the FSALs (except FSAL_PSEUDO and FSAL_MDCACHE*) are linked
with libganesha_nfsd.so, (and ${SYSTEM_LIBS} and ${LTTNG_LIBS}) for
resolution of various calls including calls back into ganesha.nfsd.
This will be required if, e.g., Fedora, decides to require default
linking with '-z defs'. They tried to do it for F29 but retreated
when they found that too many applications broke. In the end this may
all be just "truth and beauty."

*FSAL_PSEUDO and FSAL_MDCACHE are special; they are included in the
DSO so do not need to be (and can't be) linked with the DSO.

There might be more cmake-ish ways to create the various libs, e.g.
as OBJECTs instead of STATICs but this seems to require substantially
more changes everywhere. There might also be a better way to link with
libganesha_nfsd.so but I don't know cmake well enough to determine what
that might be.

A couple other things that could be done are
1) link libganesha_nfsd.so with a .sym file to explicitly control which
symbols are visible to the FSALs.
2) don't link ganesha.nfsd with -bdynamic as it's no longer needed

Change-Id: I6ddb691845e45094b7c636f3d5530a595c13acb2
Signed-off-by: Kaleb S. KEITHLEY <kkeithle@redhat.com>