Update patch set 2
Abandoned

We are NOT really caching the full reply when sa_cachthis is NOT set. There would be a memory leak with this patch. Since we change that status to NFS4ERR_RETRY_UNCACHED_REP, checking status for NFS4_OK in nfs4_op_xxx_Free() is a better option at the moment.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon