Update patch set 10
Patch Set 10:

Small revision to prepend the uuid string that we pass to cephfs with "ganesha-". This is just to improve uniqueness, as I was concerned we could end up with uuid collisions just using the nodeid. For instance, we could end up running samba in the same cluster and we'd want our uuids to collide with theirs, even if they use the same nodeid scheme.

Patch-set: 10
Reviewer: Jeff Layton <1008909@d5d70762-12d0-45a1-890d-524b12d3f735>