Update patch set 1
Patch Set 1:

I have mixed feelings about this given how verbose our logs are and how easily confused people can be; I bet most people won't notice for a while.
Haven't opened the code so lazy-me speaking again but in case we're replacing an option by another like 9p would it be possible to point to the new option programmatically so that it would print a warning but still do the correct thing over the span of one transition release?

The problem being it's not just the option name but the whole context that changes, I'm not even sure how the nested blocks + field structure would be described... And some option might just not have equivalents anymore at some point but I guess we don't deprecate these often..

Patch-set: 1
CC: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>