Update patch set 1
Patch Set 1: Code-Review-1

Actually, the compiler is probably correct.  The inline is really a switch, and the FREE case doesn't touch the variable, while the ENCODE case doesn't 
initialize the variable.  So it's a hidden initialization failure.

The following code assumes that this is a DECODE.  That code is wrong.

Patch-set: 1
Reviewer: William Allen Simpson <1005377@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1