NFSv4: avoid delays when client recalls delegations
We should not check for conflicting delegations if the client already has one.
If the client opens with CLAIM_DELEG, usually done when recalling a delegation
according to the RFC, we should not delay it. Otherwise it adds a lot of
overhead to some workloads (configure scripts, compilation) due to the use of
sillyrename by the Linux NFS client (unlinking an open file).

Change-Id: I6ac0ddc3966374c2f0552080b8b8ff7b0b2daa27
Signed-off-by: Fatih Acar <fatih.acar@gandi.net>