Update patch set 1
Patch Set 1:

Note that SIGHUP never causes unexport, so this would require a DBUS unexport. I'm not quite sure this is the best way to handle the situation, but we do rescan the filesystems on VFS or GPFS create_export (we should code the remove_fs in GPFS unexport also - and maybe we should rework release_posix_filesystem and use that instead of unclaim_fs (it will do unclaim_fs, remove_fs, and free_fs).

If we do that, the glist_for_each_safe in FSAL_VFS/GPFS needs to be replaces with a glist_fist_entry since release_posix_filesystem will also release the child filesystems (part of what needs to be dealt with there is a child filesystem that belongs to a different export - something we don't handle on unexport anyway...).

Patch-set: 1
CC: Frank Filz <1003571@d5d70762-12d0-45a1-890d-524b12d3f735>