cmake: always build in liburcu
There is some infrastructure in liburcu for handling things locklessly
that could be of great use to ganesha. That library is available on
both Linux and FreeBSD, so there should be no reason we can't use it
with ganesha.

Change cmake to always search for liburcu and fail to configure if it
isn't present.

Change-Id: I7404115a762dc4d4bb8547f0fe7dcde2d9f3720e
Signed-off-by: Jeff Layton <jlayton@redhat.com>