rados_recov: convert rados_recov_oid to be RCU-managed gsh_refstr
In the big scheme of things, the rados_recov_oid changes very
infrequently, and only in response to a different server requesting a
grace period.

Change the rados_recov_oid to be managed with an atomic refcount and
RCU. The general pattern is to take the rcu_read_lock and atomically
bump the refcount and then release it. Once we are finished with the
string we drop the refcount.

When we need to change the string, we swap the old object with the new
with an xchg operation, and then call synchronize_rcu to wait for any
users of the old pointer to complete.

This allows us to fetch and update the thing locklessly but safely.

Change-Id: I6df9ce3f0f562b8d760d4c02311528bbe635d17b
Signed-off-by: Jeff Layton <jlayton@redhat.com>