Update patch set 1
Patch Set 1:

It's fine to do this in a fsal helper, where such a helper exists.  This one changes semantics slightly, though, for NFSv4, in that it checks on all handles (rather than just the anonymous one).  I'm not sure if that's a problem or not.

Patch-set: 1
Reviewer: Daniel Gryniewicz <1004087@d5d70762-12d0-45a1-890d-524b12d3f735>