Update patch set 1
Patch Set 1:

(1 comment)

9P doesn't have any sort of replay so making it async *should* be trivial...  In theory.
In practice, this won't be easy mostly because of the way we preallocate and inconditionally send reply kind of sucks (e.g. _9p_write is called with a `preply` pointer to the buffer that will be sent if _9p_write returns success)

The least change approach to make that really async would be to replace that return value from int to an enum with Ok, Ok_async (no send), Error or something and have the write cb get a new reply buffer when it needs to send something instead.
In tcp the reply buffer is allocated on the stack; with rdma it needs registering so they're reused in a list and there's a mutex+cond to grab one safely, so making something generic is not as easy as it sounds... I guess that could be shoved into its own function and made to allocate on the heap for tcp...

Well, the real least change approach would be some sort of mutex/cond in the cb to wake the protocol thread, I guess that won't be worse than right now.

I'm curious though, sorry if this been said before as I haven't been following, the plan is to let the write2/read2 op decide if it wants to inline the IO or not? I believe for small (<128k?) IO it'd be faster to not do async in general..

Patch-set: 1
Reviewer: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>