Update patch set 1
Patch Set 1:

(1 comment)

Not sure I like the design of using an external cond/mutex; the whole async_process_data struct could be allocated on the stack in fsal_read/write everytime it's needed.
Making everyone provide a mutex/cond can lead to one being used multiple times and stuff like that which don't really make sense for this (but still work mostly with a broadcast, as long as the actual data struct is different for the 'done' flag, but I don't see what one would gain from that)

Did you have a reason for that I'm missing? the mutex/cond aren't any bigger than the pointers to them, so it'd be just as well to embed them in as long as it's initialized properly.

Patch-set: 1
CC: Dominique Martinet <1003556@d5d70762-12d0-45a1-890d-524b12d3f735>